+++ config.yml
@@ -1,0 +1,9 @@
+batch_size: 64
+learning_rate: 0.001
+epochs: 10
+num_partitions: 5
+
+model_name: enet0
+device: cpu
+
+data_path: CIFAR10/
