+++ main_distributed.py
@@ -1,7 +1,6 @@
 import os
 import torch
 import torch.nn as nn
-import torch.optim as optim
 from torchvision import datasets, transforms
 from torch.utils.data import DataLoader, Subset
 import multiprocessing
@@ -16,69 +15,21 @@ from networks.resnet50 import Resnet50
 from networks.resnet18 import Resnet18
 from dataloader.cifar10_dataset import CIFAR10Dataset
 from dataloader.dataloader import get_data_loaders
-from train.train import train_distributed
 from train.val import val
 from test.test import test
 
-# Pytorch DDP
-# https://pytorch.org/tutorials/intermediate/ddp_tutorial.html
+from utils import train_model, get_model_parameters, average_model_parameters, average_model_gradients, apply_averaged_parameters_and_gradients
 
 # Hyperparameters (can use CLI)
 batch_size = 64
 learning_rate = 0.001
 epochs = 10
-model_name = 'enet0'
+model_name = 'SimpleCNN'
 
 # device = torch.device("mps" if torch.backends.mps.is_available() else "cpu")
 device = 'cpu'
 print(f"Using device: {device}")
 # print("Number of CPUs being used", multiprocessing.cpu_count())
-
-# each process will have their own
-def train_model(model, train_loader, queue, epoch):
-    # print("Number of CPUs being used", multiprocessing.cpu_count())
-    # Model & loss & optimizer
-    criterion = nn.CrossEntropyLoss()
-    optimizer = optim.Adam(model.parameters(), lr=learning_rate)
-
-    print('training')
-    gradients = train_distributed(model, device, train_loader, optimizer, criterion, epoch)
-    queue.put(gradients)
-    
-
-def get_model_parameters(model):
-    """ Extract parameters from a single model. """
-    parameters = {name: param.clone().detach() for name, param in model.named_parameters()}
-    return parameters
-
-
-def average_model_parameters(model_parameters_list):
-    """ Average the parameters of models in a list. """
-    avg_parameters = {}
-    for key in model_parameters_list[0].keys():
-        # Stack the same parameter from each model and then take the mean
-        avg_parameters[key] = torch.stack([params[key] for params in model_parameters_list]).mean(dim=0)
-    return avg_parameters
-
-
-def average_model_gradients(gradient_list):
-    """ Average the gradients of models in a list. """
-    avg_gradients = {}
-    for key in gradient_list[0].keys():
-        # Stack the same gradient from each model and then take the mean
-        avg_gradients[key] = torch.stack([grads[key] for grads in gradient_list]).mean(dim=0)
-    return avg_gradients
-
-
-def apply_averaged_parameters_and_gradients(model, avg_parameters, avg_gradients):
-    """ Apply averaged parameters and gradients to a model. """
-    with torch.no_grad():
-        for name, param in model.named_parameters():
-            if name in avg_parameters:
-                param.copy_(avg_parameters[name])
-            if param.grad is not None and name in avg_gradients:
-                param.grad.copy_(avg_gradients[name])
-
 
 def main():
     start_time = time.time()
@@ -147,7 +98,7 @@ def main():
         for i in range(num_partitions):
             queue = Queue()
             queues.append(queue)
-            p = Process(target=train_model, args=(models[i], train_loaders[i], queue, epoch))
+            p = Process(target=train_model, args=(models[i], train_loaders[i], queue, epoch, learning_rate, device))
             p.start()
             processes.append(p)
